requirements.txt
absl-py
accelerate
accelerator
aiofiles
aiohttp
aiosignal
altair
anthropic
anyio
appdirs
astroid
asttokens
async-timeout
attrs
backcall
backports.functools-lru-cache
beautifulsoup4
bitsandbytes
black
botorch
bottle
Brotli
cachetools
certifi
charset-normalizer
click
cmake
colorama
coloredlogs
comm
contextlib2
contourpy
cycler
datasets
debugpy
decorator
deepspeed
diffusers
dill
distro
docker-pycreds
docopt
einops
entrypoints
evaluate
exceptiongroup
executing
fastapi
ffmpy
filelock
fire
fonttools
frozenlist
fschat
fsspec
ftfy
gdown
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-auth
google-auth-oauthlib
google-generativeai
googleapis-common-protos
gpytorch
gradio
gradio_client
grpcio
grpcio-status
h11
hjson
httpcore
httpx
huggingface-hub
humanfriendly
idna
importlib-metadata
inflate64
ipykernel
ipython
isort
jaxtyping
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter-client
jupyter_core
kiwisolver
lazy-object-proxy
linear-operator
linkify-it-py
lit
llama-recipes
loralib
Markdown
markdown-it-py
markdown2
MarkupSafe
matplotlib
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
mediapy
ml-collections
mpmath
multidict
multipledispatch
multiprocess
multivolumefile
mypy-extensions
nest-asyncio
networkx
nh3
ninja
nltk
num2words
numpy
nvidia-cublas-cu11
nvidia-cublas-cu12
nvidia-cuda-cupti-cu11
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu11
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu11
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu11
nvidia-cudnn-cu12
nvidia-cufft-cu11
nvidia-cufft-cu12
nvidia-curand-cu11
nvidia-curand-cu12
nvidia-cusolver-cu11
nvidia-cusolver-cu12
nvidia-cusparse-cu11
nvidia-cusparse-cu12
nvidia-nccl-cu11
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu11
nvidia-nvtx-cu12
oauthlib
openai
opt-einsum
optimum
orjson
packaging
pandas
parso
pathspec
pathtools
peft
pexpect
pickleshare
Pillow
pip
platformdirs
pretty-errors
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
py-cpuinfo
py7zr
pyarrow
pyasn1
pyasn1-modules
pybcj
pycodestyle
pycryptodomex
pydantic
pydocstyle
pydub
Pygments
pylint
pynvml
pyparsing
pyppmd
pyro-api
pyro-ppl
PySocks
python-dateutil
python-dotenv
python-multipart
pytorch-ranger
pytz
PyYAML
pyzmq
pyzstd
referencing
regex
requests
requests-oauthlib
responses
rich
rouge
rpds-py
rsa
safetensors
scikit-learn
scipy
seaborn
semantic-version
sentence-transformers
sentencepiece
sentry-sdk
setproctitle
setuptools
shortuuid
six
smmap
sniffio
snowballstemmer
soupsieve
stack-data
starlette
svgwrite
sympy
tabulate
tenacity
tensorboard
tensorboard-data-server
tensorboardX
termcolor
texttable
threadpoolctl
tiktoken
tokenize-rt
tokenizers
toml
tomli
tomlkit
toolz
torch
torch-optimizer
torchaudio
torchvision
tornado
tqdm
traitlets
transformers
triton
typeguard
typing_extensions
tzdata
uc-micro-py
urllib3
uvicorn
waitress
wandb
wavedrom
wcwidth
websockets
Werkzeug
wheel
wrapt
xxhash
yacs
yarl
zipp
transformers
colorama
openai
SentencePiece